requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
paddlepaddle>=2.4.2
paddleocr>=2.7.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0